Criando métodos para autenticação de rotas

+++ server/src/routes/memories.ts
@@ -4,8 +4,15 @@ import { prisma } from '../lib/prisma'
 import {z} from 'zod'
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get('/memories', async () => {
+  app.addHook('preHandler', async (request) => {
+    await request.jwtVerify()
+  })
+  app.get('/memories', async (request) => {
+    await request.jwtVerify()
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: {
         createdAt: 'asc'
       }
@@ -19,7 +26,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
     })
   })
 
-  app.get('/memories/:id', async (request) => {
+  app.get('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -30,6 +37,10 @@ export async function memoriesRoutes(app: FastifyInstance) {
         id,
       },
     })
+
+    if(!memory.isPublic && memory.userId !== request.user.sub){
+      return reply.status(401).send() 
+    }
     return memory
   })
 
@@ -46,14 +57,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: "8c78c335-ce74-4cb6-9f22-153d17f87086",
+        userId: request.user.sub,
       }
     })
 
     return memory
   })
 
-  app.put('/updateMemories/:id', async (request) => {
+  app.put('/updateMemories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -64,7 +75,16 @@ export async function memoriesRoutes(app: FastifyInstance) {
       isPublic: z.coerce.boolean().default(false),
     })
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
-    const memory = await prisma.memory.update({
+
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      }
+    })
+    if(memory.userId !== request.user.sub){
+      return reply.status(401).send() 
+    }
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -74,12 +94,19 @@ export async function memoriesRoutes(app: FastifyInstance) {
     })
     return memory
   })
-  app.delete('/deleteMemories/:id', async (request) => {
+  app.delete('/deleteMemories/:id', async (request, reply) => {
      const paramsSchema = z.object({
       id: z.string().uuid(),
     })
     const { id } = paramsSchema.parse(request.params)
-
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      }
+    })
+    if(memory.userId !== request.user.sub){
+      return reply.status(401).send() 
+    }
     await prisma.memory.delete({
       where: {
         id,
